Replace custom hasher pool with sync.Pool
